attempt to fix error

+++ chip8.hpp
@@ -160,7 +160,7 @@ private:
 
 	std::default_random_engine randGen;
 	
-	std::uniform_int_distribution<int> randByte{0, 255};	
+	std::uniform_int_distribution<unsigned short> randByte{0, 255};	
 
 	typedef void (Chip8::*Chip8Func)();
 	Chip8Func table[0xF + 1];
